+++ src/components/accordion/AccordionComponent.vue
@@ -78,7 +78,7 @@ export default {
   display: flex;
   flex-direction: row;
   background-color: $neutral-color-50;
-  border: 1px solid $neutral-color-400;
+  border: 1px solid $neutral-color-100;
   padding: 10px;
   border-radius: 5px;
   cursor: pointer;
@@ -100,7 +100,7 @@ export default {
 
 .gts-accordion-content {
   padding: 10px;
-  border: 1px solid $neutral-color-400;
+  border: 1px solid $neutral-color-100;
   border-top: 0;
 }
 </style>